Create drawingOnCanvasHelper file to separate code from the HangmanCanvas file
Create drawOnCanvas function in the helper to handle the drawing on canvas.

Use the newly created context in the hangmanCanvas component.
